Renamed BlockDestroyedEvent into BlockDropsEvent and moved it into its own class.
